Given a Professional File Structure for our Redux app
