+++ 18-Redux-intro/src/index.js
@@ -2,8 +2,10 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import './index.css'
 import App from './App'
+import store from './store'
 
-import './store'
+store.dispatch({ type: 'account/deposit', payload: 250 })
+console.log(store.getState())
 
 const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(
